extract the http part of querying the NCUA for CreditUnion::Details into CreditUnion::DetailsClient
